extract classes to files
